✨ feat: #13 Gridコンポーネント内で、エディタ時のみ表示するコンポーネントを分割
